Build cleaner.

+++ services/about/java/samples/java/AboutServerSample/src/org/alljoyn/about/sample/AboutServerSample.java
@@ -327,7 +327,7 @@ public class AboutServerSample {
         PropertyStore impl=new PropertyStoreImpl(data);
 
         try{
-            AboutServiceImpl.getInstance().startAboutServer((short)CONTACT_PORT, impl, mBus);
+            AboutServiceImpl.getInstance().startAboutServer(CONTACT_PORT, impl, mBus);
         }catch (Exception e){
             e.printStackTrace();
         }
